init: repo-search: initial revision
Add code for searching through the current git repo for a file matching
the given glob pattern. A new window is opened with all matching files
and the file to open may be selected with RET.

The major mode in the new buffer is help-mode, so the normal help
bindings apply (including q to quit without opening a new file).

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>